// PolicyDropDownControl/policy_dropdown/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    apiUrl: ComponentFramework.PropertyTypes.StringProperty;
    selectedValue: ComponentFramework.PropertyTypes.StringProperty;
}
export interface IOutputs {
    selectedValue?: string;
}